my spheres so far
uploaded some ideas for the spheres
press: 1 for space sphere song, 2 for city sphere song, 3 for jungle
sphere song

make sure your irrkrang library is working, let me know if it works or
if it doesnt. i also added another song (wonderful star walk is wonder),
that can be our music when we are not in the sphere. let me know what
you think.